refactor(identity): the fold splits per event, and one test file becomes four
The review's findings, all five of them:

- `reduceIdentifier` was one 80-line switch well past the complexity limit.
  It is now a dispatcher over seven named folds — `foldAttached`,
  `foldVerified`, `foldDeadEnded`, `foldPrimaryChanged`, `foldDetached`,
  `foldErased` — each holding the guards for its own event and nothing else.
  Every branch is the same branch it was; the suite that pins the two folds
  against each other is what says so.
- The 676-line test file becomes four, along the boundaries it already had:
  routing, folding, the two histories one head cannot see, and the facts a
  command states. The fixtures move to `__tests__/support/identifier-facts.ts`,
  which is where this repo keeps them.
- Those fixtures took positional arguments. `attached`, `verified`,
  `detached`, `deadEnded` and `foldStream` now read the way `primaryChanged`
  already did.
- An erasure naming no identifier was stated in the spec and asserted
  nowhere. It is asserted now: the person's stream still gets the record.
- US spelling in the new ADR and the new feature file, which is what the rest
  of the corpus uses.
